Replace special chars with hyphens.
