+++ platform/functions/shared/factories/QualifyingTests/newQualifyingTestResponse.js
@@ -1,4 +1,5 @@
-
+const { getSearchMap } = require('../../search');
+ 
 module.exports = (config, firebase) => {
 
   return newQualifyingTestResponse;
@@ -54,20 +55,5 @@ module.exports = (config, firebase) => {
 
     return data;
   }
-
-  function getSearchMap(searchables) {
-    const searchMap = {};
-    const n = 3;
-    searchables.forEach(searchable => {
-      if (searchable) {
-        const src = searchable.toLowerCase();
-        for (let i = 0, len = src.length - n; i <= len; ++i) {
-          searchMap[src.substring(i, i + n)] = true;
-        }
-      }
-    });
-    return searchMap;
-  }
-
 };
